Modification of the update condition of quadrature points
The update is now only done when one of the components of the strain
tensor exceeds the 'strain_perturbation' used in LAMMPS to compute
the tangent stiffness tensor, in which we assume a linear behaviour.

The workflow detects that a LAMMPS update has to be run for a specific
gauss point if there exists a file with .strain extension, for the gauss
point at the current time-step iteration.